Shadowcaster upgrades, few extra fixes and a issue
-Translated all the Shadowcaster upgrade cards, and added the "ship" tag to the ShadowCaster as the slot was missing.
-Fixed a typo on the "ship:" tag on the Protectorate Starfighter pilots ("Caza estelar...." when should be "Caza Estelar...".), as the pilots aren't showing up right now.
-Added "ship:" tag to the /sf and Sabine Masterpiece titles, as they were missing.

Now, the issue: ARC-170 has disappeared from the builder. Looking the file, it occurs to me that may be caused me adding it to the "renameship" list... something that seems I shouldn't have done due being the same name. I mean, yesterday I didn't realize that the few ships with the same name on both languages (both YT and the YV-666) aren't on the list. I guess this may be the cause of the Travis error yesterday.

I've deleted the line to see if Travis gives me the middle finger again or not. Either way, I suck  :(