FAQs
Updated the cards with the changes made to them due erratas and nerfs. 

This is something I used to do time ago, when I started translating this, but I stopped as the official Spanish version of the FAQs was usually released with a 8-10 months delay. But as the next Tatooine Open will use this builder as *the* builder, I thought I should update everything to avoid problems.