C-ROC
Translated all the cards from the C-ROC expansion.

I have a doubt with the "Quick-release Cargo Locks" tho. Usually I would add the Ship: tag so it won't go missing due the translated ship name, but in this case is an upgrade for *two* ships, C-ROC and GR-75. How must I put this on these cases? Two regular Ship: tags, one for each ship?

Right now I didn't add an tag on it.