Allow printing of dials in fancy view
Adding a option to include maneuvers to the print.

Currently the fancy-view (on screen) will always show the maneuvers, but
that could obviously be turned off easily.

There is a checkbox to enable printing the maneuvers of the included
ships. The maneuvers are inserted next to the pilot ability, if any
exists, or above the upgrades otherwise. They will also be printed nicely
in black-white mode, including a highlight of maneuvers with not-default
difficulty (e.g. R4-Astromech).