Introducing display_names
This introduces display_names for all kind of stuff. We will now only
translate display names, so logic is not affected anymore. E.g. we do not
specify each translated shipname for each pilot or title. Second edition
only works out of the box. The collection check is able to do it's work
regardless of language, ...