Improve xws import/export
We now follow xws 2.0.0 specs on our imports (we no longer rely on beeing
provided with a ship for each pilot, but should be able to work out the
ship from the pilot id).

We allow others to still use name instead of id, without breaking our
import.

Our export is exporting ship, id and name for all pilots. This might allow
importing our exports in other squad builder, which did not fully
implement xws 2.0.0 yet. According to the spec, one only has to include
the id. We do that for our minimal-xws, used to create barcodes.