Rework translation order
I defined a translation order: Try to find a translation in the selected
language, use english if none is available. For card texts use polish if
english isn't available as well (as some cards might be spoilered in
polish first).

This will allow, to just skip cards in language files, that do not have an
official translation yet - and will automatically display an available
translation in english instead of having a blank text field.