Change Nashtah Pup to initiative 0
The NP does not have an initiative - it inherits the initiative of the Hounds Tooth that it escapes from.  For this reason, setting the initiative to an non-standard value (like zero) would set it apart more than it having an initiative of 1.  Also, this would make less noise in my data validator! :)