Update Quickbuilds: Add Hera
this required changeing the linking of wingmates from the linking of
usual two-ship quickbuild card buddies. Hera has a phantom as buddy, and
a bunch of A-Wings as wingmates.

I've done some tests and everything seemed to work rather fine

open todo (was also previous open): check if number of wingmates is
acually within the allowed range