Code Cleanup XWS
There is some not existing html and UI (the xws used to be a dedicated
section at the bottom of the page, but has been moved to the save/load
area (for xws import) and the print/view area (for xws export). A bunch
of code was left behind.

Also fixes a bug where a @ was too much