Testing is still on 1.0 and not working
The last release of CasperJS, which was used as testing framework for
YASB, is from 2017. Though it still installed on my system, it failed
most of its self-tests and did not even attempt to run any of the tests
within YASB (`npm test` failed as well as a direct call to
`casperjs test tests/test*.coffee`). Given that the test cases themself
are probably mostly obsolete as well (they are likely untouched since
YASB 1.0), I don't see any point of having parts of the testing
framework around anymore. If we decide to introduce testing back to our
developement stack, we have to start from scratch anyway.